+++ packages/CVMTests/src/package-test.ts
@@ -4,6 +4,8 @@ import * as DG from 'datagrok-api/dg';
 import './shell/ml';
 import './scripts/scripts_tests';
 import './celery/celery_tests';
+import './celery/node_celery_tests';
+import './celery/node_celery_benchmark';
 import './docker/docker';
 import './files/files';
 // import './gui/dialogs'; To fix!