+++ packages/NodejsApiDemo/scripts/py-mean.py
@@ -1,6 +1,0 @@
-#name: PyMean
-#language: python
-#input: dataframe df
-#input: string colName
-#output: double mean_out
-mean_out = float(df[colName].mean())